# backend/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go microservice API for Quality Studio Server.",
        "title": "Quality Studio API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/konflux-ci/quality-studio"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/konflux-ci/quality-studio/blob/main/LICENSE"
        },
        "version": "v1alpha"
    },
    "host": "127.0.0.1:9898",
    "basePath": "/api/quality/",
    "paths": {
...
        "/repositories/create": {
            "post": {
                "description": "returns the Server information as a JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Github Repositories API"
                ],
                "summary": "Github repositories info",
                "parameters": [
                    {
                        "description": "repository name",
                        "name": "repository",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repositories.GitRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/delete": {
            "delete": {
                "description": "delete a given repository from a organization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Github Repositories API"
                ],
                "summary": "Github repositories info",
                "parameters": [
                    {
                        "description": "repository name",
                        "name": "repository",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repositories.GitRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/list": {
            "get": {
                "description": "returns all repository information stored in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Github Repositories API"
                ],
                "summary": "Github repositories info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.RepositoryQualityInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflows/get": {
            "get": {
                "description": "return github workflows from a given repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Github Workflows API"
                ],
                "summary": "Quality Repositories Workflow",
                "parameters": [
                    {
                        "type": "string",
                        "example": "string",
                        "description": "string example",
                        "name": "repository_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GithubWorkflows"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "repositories.GitRepositoryRequest": {
            "type": "object",
            "properties": {
                "artifacts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "git_organization": {
                    "type": "string"
                },
                "jobs": {
                    "$ref": "#/definitions/repositories.JobSpec"
                },
                "repository_name": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "repositories.GithubActionsSpec": {
            "type": "object",
            "properties": {
                "monitor": {
                    "type": "boolean"
                }
            }
        },
        "repositories.JobSpec": {
            "type": "object",
            "properties": {
                "github_actions": {
                    "$ref": "#/definitions/repositories.GithubActionsSpec"
                },
                "openshift_ci": {
                    "$ref": "#/definitions/repositories.OpenshiftCISpec"
                }
            }
        },
        "repositories.OpenshiftCISpec": {
            "type": "object",
            "properties": {
                "monitor": {
                    "type": "boolean"
                }
            }
        },
        "storage.Coverage": {
            "type": "object",
            "properties": {
                "coverage_percentage": {
                    "type": "number"
                },
                "git_organization": {
                    "type": "string"
                },
                "repository_name": {
                    "description": "RepositoryName identify an github repository",
                    "type": "string"
                }
            }
        },
        "storage.GithubWorkflows": {
            "type": "object",
            "properties": {
                "badge_url": {
                    "type": "string"
                },
                "html_url": {
                    "type": "string"
                },
                "job_url": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "workflow_name": {
                    "description": "RepositoryName identify an github repository",
                    "type": "string"
                }
            }
        },
        "storage.RepositoryQualityInfo": {
            "type": "object",
            "properties": {
                "code_coverage": {
                    "$ref": "#/definitions/storage.Coverage"
                },
                "description": {
                    "type": "string"
                },
                "git_organization": {
                    "type": "string"
                },
                "git_url": {
                    "type": "string"
                },
                "github_actions": {
                    "description": "Coverage",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.GithubWorkflows"
                    }
                },
                "repository_name": {
                    "description": "RepositoryName identify an github repository",
                    "type": "string"
                }
            }
        },
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "The error message.\nRequired: true",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The error message.\nRequired: false",
                    "type": "integer"
                }
            }
        },
        "types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "The message.\nRequired: true",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The error message.\nRequired: false",
                    "type": "integer"
                }
            }
        },
...
    }
}